Fixed platform-dependent bug that was preventing the install scripts from running.

+++ Source/CDRSlowTestReporter.m
@@ -100,12 +100,12 @@
     NSArray *sortedExamplePairs = [[[examplePairs sortedArrayUsingSelector:@selector(compare:)] reverseObjectEnumerator] allObjects];    
     sortedExamplePairs = [sortedExamplePairs subarrayWithRange:NSMakeRange(0, MIN(numberOfResultsToShow, sortedExamplePairs.count))];
     
-    printf("\n%lu Slowest Tests\n\n", sortedExamplePairs.count);
+    printf("\n%d Slowest Tests\n\n", (int)sortedExamplePairs.count);
     for (RunTimeTitlePair *pair in sortedExamplePairs) {
         printf("%s\n\n", pair.formattedDescription.UTF8String);
     }
     
-    printf("\n%lu Slowest Top-Level Groups\n\n", sortedRootPairs.count);
+    printf("\n%d Slowest Top-Level Groups\n\n", (int)sortedRootPairs.count);
     for (RunTimeTitlePair *pair in sortedRootPairs) {
         printf("%s\n\n", pair.formattedDescription.UTF8String);
     }
